+++ core-ng-test/src/main/java/core/framework/module/TestKafkaConfig.java
@@ -1,15 +1,14 @@
 package core.framework.module;
 
-import core.framework.impl.kafka.Kafka;
-import core.framework.impl.module.ModuleContext;
-import core.framework.test.kafka.MockKafka;
+import core.framework.impl.kafka.MessageProducer;
+import core.framework.test.kafka.MockMessageProducer;
 
 /**
  * @author neo
  */
 public class TestKafkaConfig extends KafkaConfig {
     @Override
-    Kafka createKafka(ModuleContext context, String name) {
-        return new MockKafka();
+    MessageProducer createProducer() {
+        return new MockMessageProducer();
     }
 }