+++ core-ng/src/main/java/core/framework/impl/kafka/MessageHeaders.java
@@ -3,7 +3,7 @@ package core.framework.impl.kafka;
 /**
  * @author neo
  */
-public class KafkaHeaders {
+public final class MessageHeaders {
     public static final String HEADER_REF_ID = "refId";
     public static final String HEADER_CLIENT = "client";
     public static final String HEADER_CLIENT_IP = "clientIP";