+++ core-ng/src/main/java/core/framework/impl/kafka/MessageProducer.java
@@ -1,0 +1,10 @@
+package core.framework.impl.kafka;
+
+import org.apache.kafka.clients.producer.ProducerRecord;
+
+/**
+ * @author neo
+ */
+public interface MessageProducer {
+    void send(ProducerRecord<String, byte[]> record);
+}